Sync V4

+++ cocos/2d/CCFontFNT.cpp
@@ -617,7 +617,7 @@ FontFNT * FontFNT::create(const std::string& fntFilePath, const Vec2& imageOffse
     }
     
     FontFNT *tempFont =  new FontFNT(newConf,imageOffset);
-    tempFont->setFontSize(newConf->_fontSize);
+    tempFont->setFontSize((float)newConf->_fontSize);
     if (!tempFont)
     {
         return nullptr;
@@ -713,7 +713,7 @@ FontAtlas * FontFNT::createFontAtlas()
     
     // common height
     int originalFontSize = _configuration->_fontSize;
-    float originalLineHeight = _configuration->_commonHeight;
+    float originalLineHeight = (float)_configuration->_commonHeight;
     float factor = 0.0f;
     if (std::abs(_fontSize - originalFontSize) < FLT_EPSILON) {
         factor = 1.0f;
